"fit to parent" molview w/o explicit width, height
This option ensures that the molecule view matches its containing element width, even when that width is fluid (percentage based). Without the fit-to-parent option, the molecule container confounds layout by adjusting its own width and height to be *larger* than the original size of the containing element, and furthermore requires explicit resizing when the window size changes.

TODO: refactor the molecules view so the buttons, icons, labels, etc can be scaled (or not) independently of the overall molecules view. Currently they scale 1:1 with the view.